+++ src/codemap_details/Components/Banner.js
@@ -1,5 +1,5 @@
 import React from "react";
-import styles from "../CSS/CodeMapBanner.module.css";
+import styles from "../CSS/Banner.module.css";
 
 export default function Banner() {
   return <React.Fragment />;
